feat(filter): add higher-order lettable version of filter
