+++ src/operators/index.ts
@@ -1,3 +1,5 @@
 export { OperatorFunction } from './OperatorFunction';
+export { filter } from './filter';
+
 export { map } from './map';
 export { mergeMap } from './mergeMap';
